[fe] [refactor]: 컨벤션에 따른 파일명 변경
